chore(security): H7+H8 — 20 nuovi golden test + feature flag + supply-chain policy
Focus della campagna: chiudere Postgres come tier production e isolare
MySQL/SQL Server come feature opt-in in attesa di Fase 2.

H7 — Coverage sui path plan-based del provider (Provider::read/inspect/write)
che erano <10% coperti pur essendo vivi via CLI e via consumer futuri:

- tests/golden_inspect_catalog.rs (5 live): DatabaseListCatalogs,
  DatabaseListSchemas (documenta finding: NON filtra system schemas),
  DatabaseListObjects, DatabaseDescribeObject su spatial_ref_sys +
  tabella con tipi eterogenei.
- tests/golden_read_arrow.rs (6 live): read Arrow BatchStream su
  spatial_ref_sys, schema eterogeneo (jsonb/numeric/timestamptz),
  projection reduction, row_limit, null misti bit-perfect, e
  documentazione del finding: BatchStream::next_batch() del trait
  pubblico NON è cancel-aware (esiste next_batch_with_cancellation
  ma è pub-crate).
- tests/golden_write_bulk.rs (4 live): Provider::prepare_write + write
  in WriteMode::Create/Append/Replace, verifica outcome.rows.confirmed,
  errore classificato quando target inesistente.

Totale H7: +15 test live golden (unignored: 601 → 616 + 4 defer live).
Trovati 2 finding reali (list_schemas raw, BatchStream non cancel-aware)
documentati nei test e da fixare in v0.2.

H8 — Isolamento e supply-chain hygiene:

- Feature flag CLI (`crates/plenora-database-cli/Cargo.toml`):
    default = ["postgres"]
    postgres/mysql/sqlserver = opt-in via dep:
    full = tutti
  Dipendenze plenora-db-mysql / plenora-db-sqlserver marcate optional.
  Tutti gli import + fn helper + varianti enum + dispatcher branch +
  test inline cfg-gated con #[cfg(feature = "...")].
  Effetto: `cargo build -p plenora-database-cli` di default NON compila
  tiberius/mysql_async — CVE transitive non entrano nel binario PFM.

- deny.toml (cargo-deny policy strutturata):
    advisories = deny, con 3 eccezioni documentate:
      * RUSTSEC-2026-0098/0099/0104 → rustls-webpki 0.101.7 via
        tiberius 0.12.3 (tokio-rustls 0.24 → rustls 0.21). Non entra
        in default build. Rimuovere quando tiberius bump rustls ≥ 0.23.
      * RUSTSEC-2025-0134 → rustls-pemfile unmaintained. Info-only,
        codice stabile, migrazione opzionale. Rivedere 2026-Q3.
    licenses = allowlist ristretta (MIT/Apache-2.0/BSD-*/ISC/Zlib/
    Unicode-3.0/CC0-1.0/Unlicense/CDLA-Permissive-2.0 per webpki-roots).
    bans = wildcards deny esterne, allow-wildcard-paths per workspace,
    multiple-versions warn (arrow-*, syn hanno doppi genuini).
    sources = solo crates.io.
  Risultato: `cargo deny check` → advisories ok, bans ok, licenses ok,
  sources ok.

- Workspace license.workspace = "Apache-2.0" per soddisfare cargo-deny
  license check sui crate interni (publish=false comunque).

- Workspace default-members esclude i provider driver non-Postgres:
  `cargo build/test/clippy` senza `-p` opera solo su core+Postgres+CLI.
  MySQL e SQL Server restano compilabili on-demand.

Fix collaterali:
- transaction/params.rs: fn encode(v) → encode(&v) + call sites (clippy).
- tests/error_protocol.rs: 8 test che esercitavano paths MySQL/SQL
  Server gate con #[cfg(all(feature = "mysql", feature = "sqlserver"))]
  così restano validi solo in `--features full`.
- tests/live_f4.rs: cast u64→usize → asserzione a u64.
- Golden tests: crate-level #![allow] per pedantic style-only lint
  irrilevanti sui test di integrazione (approx_constant, casts,
  uninlined format, etc).

Contatori finali:
- Test default build: 582 verdi (Postgres-only)
- Test --features full: 594 verdi
- clippy pedantic+nursery: 0 warning su default e full
- cargo deny check: ✓ tutti i 4 gate

Effetto per il PFM v1: il binario distribuito `plenora-database` con
default features è certificato Postgres, senza CVE transitive nel path
attivo. Fase 2 sbloccherà MySQL/SQL Server quando gli upstream saranno
bonificati.

+++ crates/plenora-database-cli/tests/live_f4.rs
@@ -59,7 +59,7 @@ fn run_expect_error(args: &[&str]) -> Value {
     })
 }
 
-#[ignore]
+#[ignore = "live: richiede Postgres su dataflow-postgres"]
 #[test]
 fn f4_inspect_database_returns_expected_metadata_shape() {
     let out = run(&["inspect-database", "PG_DSN"]);
@@ -80,13 +80,13 @@ fn f4_inspect_database_returns_expected_metadata_shape() {
     );
 }
 
-#[ignore]
+#[ignore = "live: richiede Postgres su dataflow-postgres"]
 #[test]
 fn f4_inspect_schemas_lists_user_schemas() {
     let out = run(&["inspect-schemas", "PG_DSN"]);
     let count = out["count"].as_u64().expect("count u64");
     let schemas = out["schemas"].as_array().expect("schemas array");
-    assert_eq!(count as usize, schemas.len(), "count/schemas mismatch");
+    assert_eq!(count, schemas.len() as u64, "count/schemas mismatch");
     // 'public' deve esserci sempre; pg_catalog non deve.
     assert!(
         schemas.iter().any(|s| s["name"] == "public"),
@@ -101,7 +101,7 @@ fn f4_inspect_schemas_lists_user_schemas() {
     }
 }
 
-#[ignore]
+#[ignore = "live: richiede Postgres su dataflow-postgres"]
 #[test]
 fn f4_inspect_tables_lists_relations_with_size_and_row_estimate() {
     let out = run(&["inspect-tables", "PG_DSN", "public"]);
@@ -116,7 +116,7 @@ fn f4_inspect_tables_lists_relations_with_size_and_row_estimate() {
     }
 }
 
-#[ignore]
+#[ignore = "live: richiede Postgres su dataflow-postgres"]
 #[test]
 fn f4_diagnose_reports_healthy_when_probes_pass() {
     let out = run(&["diagnose", "PG_DSN"]);
@@ -146,7 +146,7 @@ fn f4_diagnose_reports_healthy_when_probes_pass() {
     assert!(out["profiles"]["PFM_GIS_V1"]["status"].is_string());
 }
 
-#[ignore]
+#[ignore = "live: richiede Postgres su dataflow-postgres"]
 #[test]
 fn f4_benchmark_write_requires_explicit_gate() {
     // Senza --allow-write-tests → invalid_plan
@@ -159,7 +159,7 @@ fn f4_benchmark_write_requires_explicit_gate() {
         .contains("--allow-write-tests"));
 }
 
-#[ignore]
+#[ignore = "live: richiede Postgres su dataflow-postgres"]
 #[test]
 fn f4_benchmark_write_reports_throughput_and_percentiles() {
     let out = run(&[
@@ -182,7 +182,7 @@ fn f4_benchmark_write_reports_throughput_and_percentiles() {
     assert!(p50 <= p95 && p95 <= p99, "percentili non monotoni: p50={p50} p95={p95} p99={p99}");
 }
 
-#[ignore]
+#[ignore = "live: richiede Postgres su dataflow-postgres"]
 #[test]
 fn f4_test_concurrency_reports_winner_and_loser_correctly() {
     let out = run(&[
@@ -195,7 +195,7 @@ fn f4_test_concurrency_reports_winner_and_loser_correctly() {
     assert_eq!(out["loser_error_category"], "ConcurrentModification");
 }
 
-#[ignore]
+#[ignore = "live: richiede Postgres su dataflow-postgres"]
 #[test]
 fn f4_profile_check_returns_pass_for_application_oltp_v1() {
     let out = run(&["profile-check", "PG_DSN", "APPLICATION_OLTP_V1"]);
@@ -205,7 +205,7 @@ fn f4_profile_check_returns_pass_for_application_oltp_v1() {
     assert_eq!(out["failed"].as_array().map(Vec::len), Some(0));
 }
 
-#[ignore]
+#[ignore = "live: richiede Postgres su dataflow-postgres"]
 #[test]
 fn f4_format_junit_wraps_ok_status_as_system_out() {
     // JUnit non è JSON: parse manuale minimale.
@@ -226,7 +226,7 @@ fn f4_format_junit_wraps_ok_status_as_system_out() {
     assert!(stdout.contains("<system-out>"));
 }
 
-#[ignore]
+#[ignore = "live: richiede Postgres su dataflow-postgres"]
 #[test]
 fn f4_format_markdown_renders_title_and_bullets() {
     let output = Command::new(BIN)
